requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
boto==2.49.0
boto3==1.16.18
botocore==1.19.18
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
dill==0.3.3
discord.py==1.5.1
distlib==0.3.1
filelock==3.0.12
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.11.21
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
ibm-cloud-sdk-core==1.7.3
ibm-watson==4.7.1
idna==2.10
isort==5.6.4
jmespath==0.10.0
kiwisolver==1.2.0
klepto==0.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.2
Pillow==7.2.0
pox==0.2.9
pycodestyle==2.6.0
PyJWT==1.7.1
pykakasi==2.0.4
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.0
rfc3986==1.4.0
s3transfer==0.3.3
selenium==3.141.0
six==1.14.0
sniffio==1.2.0
toml==0.10.1
urllib3==1.25.10
virtualenv==20.0.28
websocket-client==0.48.0
wrapt==1.12.1
yarl==1.5.1